Fix problem running test executables when compiled with newer compilers.
With newer systems there is a link problem when linking the executable against ROOT libraries. Somehow libraries
which are needed due to a dependency of another library (defined in the link command) are not linked correctly.
This results in errors at runtime since the executable can't load all needed libraries and crashes. Up to now it
was not possible to find out what has changed. To bypass the problem for testing we create a shell script which
sets the proper runtime environment and then executes the executable. The proper solution is to properly link the
executables such that all libraries are founs at runtime.