Define accepted and expected overlaps.
For the time beeing we have to live with overlöaps between the magnet and the rich or the magnet and the
first much absorber. These overlaps are defined as expected and not counted in the overall number of overlaps.
If the geometry has only such expected overlaps the test will pass.