Compile Vc always.
Do not use Vc if it is delivered with ROOT. The version included in ROOT is rather old and not
build in any case. Up to now we checked if ROOT includes Vc and only build an internal version
if not. Now build Vc in any case and use the internal one.