Fix problem when computer isn't connected to the internet during compilation.
KFParticle and Vc were defined as external projects in such a way hat during compilation it was checked if 
the external repositories contain updates. If no conection to the external repository was possible the compilation
failed at that point. Now the download was moved to the time when cmake is executed and during compile time the already
downloaded source is used which solved the above problem.
The drawback of this solution is that changes in KFParticle and/or Vc are only downloaded when running cmake.