Fix memory leak in the much digitizer.
During the digitization temporary CbmMuchSignals are constructed which are used to pile-up signals 
which takes place in the same channel at similar times such that those signals interfere with each other.
These CbmMuchSignal objects are stored in the ReadoutBuffer and in the end converted to CbmMuchDigis. 
During the reconstruction the pointers to these objects are copied around from one container to another
but finally the objects were not destroyed after creating the digis. Also create a new CbmMatch object
when creating the CbmMuchDigi. Otherwise it is not possible to destroy the CbmMuchSignal.