Fix memory leak in the rich digitizer.
A temporary map containing pointers to CbmRichDigi objects was not cleared properly such that the objects
were not destroyed. Now the objects are properly destroyed before clearing the map. 
Also create a new CbmMatch object when creating the CbmRichDigi. Otherwise it is not possible to destroy the
objects in the temporary map.