Fix reading of pluto files.
Pluto in the current form can't be compiled with ROOT6 suchn that the
Pluto generator has to be switched off. Fortunatly only some part of 
pluto is needed to allow reading of the ROOT files with the pluto 
particle information.
This part can be compiled with ROOT6 such that the needed parts have
been copied to our repository and have been adapted to work with
CbmRoot.
With this hack it is again possible to use the CbmPlutoGenerator and
read Pluto files which have been generated with ROOT5. Refs #1040 @4h