Fix a problem with wrong Vc version when using FairSoft may16
This old FairSoft version includes Vc with ROOT. This Vc version does
not support simd_cast which results in compilation errors. The problem 
is fixed by changing the order of include directories such that the Vc 
included with CbmRoot is found. Also the order of link directories is 
changed such that first our internal Vc library is found.