Fix problem with parameter loading on cbmfles01
With the linker on cbmfles01 libraries which are not needed at all even if
defined in the link command will not be linked. These libraries must be
linked because otherwise parameter containers defined inside these libs
can't be used. The containers have to be registered at the time the lib is
loaded. To solve the issue without the need to understand the linker the
library names are now passed at starttime of the parmq-server as paramters
and loaded using the ROOT functionality to load libraries at runtime. The
usage of the ROOT feature is possible since the parmq-server anyway relies
on ROOT.