Remove build directory of external package on source code updates
The external packages are build in the context of CbmRoot but are no
integral part of our build system. This resulted in the problem that
on changing external source code there was no rebuild of the code.
With the current commit this behaviour is changed such that on a
change of the source code of the external package the build directory
of this external package is removed which will result in a rebuild of
the package.