Finalyze event builder for mCBM data
Add the event builder to the build system.
Add an example macro for event building.
Add two different algorithms for event building. One is based on
a fixed time window, the other is based on a maximum time difference
between two consecutive digis.
Add the possibility to filter the events based on the number of digis
in  the differnt systems.